beginning layered graph
