+++ src/MenuHandler.java
@@ -1,0 +1,15 @@
+/**
+* Sunseeker Telemety
+*
+* Telemetry 2016
+*
+* @author Kai Gray <kai.a.gray@wmich.edu>
+*/
+
+
+package sunseeker.telemetry;
+
+
+class MenuHandler {
+	
+}
